WorkingVersion

+++ StudentManagementSystemSolution/SMSClient/Service/Students/StudentService.cs
@@ -1,27 +1,24 @@
 ﻿using Microsoft.AspNetCore.Identity;
 using Microsoft.EntityFrameworkCore;
 using SMSClient.Data.Identity;
-using SMSClient.Models;
-using SMSClient.Models.Identity;
-using SMSClient.Models.ViewModel;
 using SMSClient.Repository;
 using SMSClient.Repository.Students;
+using SMSClient.Model;
+using System.Security.Claims;
 
 namespace SMSClient.Service.Students
 {
     public class StudentService : IStudentService
     {
         private readonly IStudentRepository _studentRepository;
-        private readonly IRepository<DepartmentSemester> _departmentSemsterRepository;
         private readonly IRepository<UserInfo> _userInfoRepository;
         private readonly AspIdUsersContext _context;
 
         private readonly UserManager<ApplicationUser> _userManager;
 
-        public StudentService(IStudentRepository studentRepository, IRepository<DepartmentSemester> departmentSemsterRepository, IRepository<UserInfo> userInfoRepository, AspIdUsersContext context, UserManager<ApplicationUser> userManager)
+        public StudentService(IStudentRepository studentRepository, IRepository<UserInfo> userInfoRepository, AspIdUsersContext context, UserManager<ApplicationUser> userManager)
         {
             _studentRepository = studentRepository;
-            _departmentSemsterRepository = departmentSemsterRepository;
             _userInfoRepository = userInfoRepository;
             _context = context;
             _userManager = userManager;
@@ -45,12 +42,12 @@ namespace SMSClient.Service.Students
 
         public async Task<IEnumerable<Student>> GetStudentsUnAssignedToAnyDepartment()
         {
-            return await _studentRepository.FindAll(s => s.DepartmentId == null);
+            return await _studentRepository.FindAll(s => s.ClassId == null);
         }
 
         public async Task<IEnumerable<Student>> GetStudentsUnAssignedToAnySemester()
         {
-            var students = await _studentRepository.FindAll(s => s.SemesterId == null);
+            var students = await _studentRepository.FindAll(s => s.ClassId == null);
             return students;
         }
 
@@ -69,8 +66,15 @@ namespace SMSClient.Service.Students
                 var createdUser = await _userManager.FindByNameAsync(student.Email);
                 student.ApplicationUserId = createdUser.Id;
                 _studentRepository.Add(student);
-                var saved = await _context.SaveChangesAsync();
-                return saved > 0;
+                await _context.SaveChangesAsync();
+                var createdStudent = (await _studentRepository.FindAll(t => t.Email == student.Email)).SingleOrDefault();
+                if (createdStudent != null)
+                {
+                    var studentClaim = new Claim("student", createdStudent.Id.ToString());
+                    await _userManager.AddClaimAsync(createdUser, studentClaim);
+                    return true;
+                }
+                return false;
             }
             return false;
         }
@@ -86,19 +90,22 @@ namespace SMSClient.Service.Students
 
         public async Task<bool> DeleteStudent(Student student)
         {
+            if(student  == null) return false;
+            var user = await _userManager.FindByIdAsync(student.ApplicationUserId);
             _studentRepository.Remove(student);
-            var result = await _context.SaveChangesAsync();
-            return result > 0;
+            await _context.SaveChangesAsync();
+            var result = await _userManager.DeleteAsync(user);
+            return result.Succeeded;
         }
 
+        public async Task<Student?> GetStudentByUserId(string userId)
+        {
+            return (await _studentRepository.FindAll(s => s.ApplicationUserId == userId)).FirstOrDefault();
+        }
 
-
-
-
-
-
-
-
-      
+        public async Task<IEnumerable<Student>> GetStudentsWithClassInfo()
+        {
+            return await _studentRepository.GetStudentsWithClassInfo();
+        }
     }
 }
